+++ src/domain/usecases/egasp-validate/VariableService.js
@@ -1,4 +1,5 @@
-import { typeKeys } from "./RulesEngine";
+import { typeKeys } from "./ValueProcessor";
+
 export const eventStatuses = Object.freeze({
     ACTIVE: "ACTIVE",
     COMPLETED: "COMPLETED",
